Reforming the architecture to work by Django's standards.
Currently the architecure is unable to access Templates and Static from views.py. There is no app for the dashboard and all packages are simply stored straight on the site. I used startapp to create the dash app which will make up the energy dashboard web application. The website should be separate from the application that runs on it.